relocate default arguments and include scaling checks
