+++ tests/modified_tests/test_end_scenario_sizing.py
@@ -16,6 +16,11 @@ import numpy as np
 
 from rtctools.util import run_optimization_problem
 
+from utils_test_scaling import (
+    check_scaling,
+    create_problem_with_debug_info,
+)
+
 from utils_tests import demand_matching_test
 
 
@@ -26,10 +31,16 @@ class TestEndScenarioSizing(TestCase):
 
         base_folder = Path(run_ates.__file__).resolve().parent.parent
 
+        (
+            cls.end_scenario_sizing_highs,
+            cls.rtc_logger,
+            cls.rtc_logs_list,
+        ) = create_problem_with_debug_info(EndScenarioSizingHIGHS)
+
         # This is an optimization done over a full year with timesteps of 5 days and hour timesteps
         # for the peak day
         cls.solution = run_optimization_problem(
-            EndScenarioSizingHIGHS,
+            cls.end_scenario_sizing_highs,
             base_folder=base_folder,
             esdl_file_name="test_case_small_network_with_ates_with_buffer_all_optional.esdl",
             esdl_parser=ESDLFileParser,
@@ -65,6 +76,9 @@ class TestEndScenarioSizing(TestCase):
 
         # Check whehter the heat demand is matched
         demand_matching_test(self.solution, self.results)
+
+        # Check scaling differences and ranges in objective, matrix and rhs
+        check_scaling(self, self.rtc_logger, self.rtc_logs_list)
 
         # Check whether cyclic ates constraint is working
         for a in self.solution.energy_system_components.get("ates", []):
@@ -132,8 +146,14 @@ class TestEndScenarioSizing(TestCase):
 
         solution_unstaged = self.solution
 
+        (
+            end_scenario_sizing_highs_scaling,
+            rtc_logger,
+            rtc_logs_list,
+        ) = create_problem_with_debug_info(EndScenarioSizingHIGHS)
+
         solution_unstaged_2 = run_end_scenario_sizing(
-            EndScenarioSizingHIGHS,
+            end_scenario_sizing_highs_scaling,
             staged_pipe_optimization=False,
             base_folder=base_folder,
             esdl_file_name="test_case_small_network_with_ates_with_buffer_all_optional.esdl",
@@ -142,8 +162,14 @@ class TestEndScenarioSizing(TestCase):
             input_timeseries_file="Warmte_test.csv",
         )
 
+        (
+            end_scenario_staged_highs_scaling,
+            rtc_logger,
+            rtc_logs_list,
+        ) = create_problem_with_debug_info(EndScenarioSizingStagedHIGHS)
+
         solution_staged = run_end_scenario_sizing(
-            EndScenarioSizingStagedHIGHS,
+            end_scenario_staged_highs_scaling,
             base_folder=base_folder,
             esdl_file_name="test_case_small_network_with_ates_with_buffer_all_optional.esdl",
             esdl_parser=ESDLFileParser,
@@ -155,6 +181,13 @@ class TestEndScenarioSizing(TestCase):
 
         # Check whehter the heat demand is matched
         demand_matching_test(solution_staged, results)
+
+        # Check scaling differences and ranges in objective, matrix and rhs
+        check_scaling(
+            self,
+            rtc_logger,
+            rtc_logs_list,
+        )
 
         # Check whether cyclic ates constraint is working
         for a in solution_staged.energy_system_components.get("ates", []):
@@ -248,10 +281,16 @@ class TestEndScenarioSizing(TestCase):
                 highs_options["mip_rel_gap"] = 0.05
                 return options
 
+        (
+            test_end_scenario_sizing_discount_highs_scaling,
+            rtc_logger,
+            rtc_logs_list,
+        ) = create_problem_with_debug_info(TestEndScenarioSizingDiscountedHIGHS)
+
         # This is an optimization done over a full year with timesteps of 5 days and hour timesteps
         # for the peak day
         solution = run_optimization_problem(
-            TestEndScenarioSizingDiscountedHIGHS,
+            test_end_scenario_sizing_discount_highs_scaling,
             base_folder=base_folder,
             esdl_file_name="test_case_small_network_with_ates_with_buffer_all_optional.esdl",
             esdl_parser=ESDLFileParser,
@@ -269,6 +308,13 @@ class TestEndScenarioSizing(TestCase):
 
         # Check whether the heat demand is matched
         demand_matching_test(solution, results)
+
+        # Check scaling differences and ranges in objective, matrix and rhs
+        check_scaling(
+            self,
+            rtc_logger,
+            rtc_logs_list,
+        )
 
         # Check whether cyclic ates constraint is working
         for a in solution.energy_system_components.get("ates", []):
@@ -299,8 +345,14 @@ class TestEndScenarioSizing(TestCase):
 
         base_folder = Path(run_ates.__file__).resolve().parent.parent
 
+        (
+            test_end_scenario_sizing_loss_staged_scaling,
+            rtc_logger,
+            rtc_logs_list,
+        ) = create_problem_with_debug_info(EndScenarioSizingHeadLossStaged)
+
         solution = run_end_scenario_sizing(
-            EndScenarioSizingHeadLossStaged,
+            test_end_scenario_sizing_loss_staged_scaling,
             base_folder=base_folder,
             esdl_file_name="test_case_small_network_with_ates_with_buffer_all_optional.esdl",
             esdl_parser=ESDLFileParser,
@@ -311,6 +363,13 @@ class TestEndScenarioSizing(TestCase):
         results = solution.extract_results()
 
         demand_matching_test(solution, results)
+
+        # Check scaling differences and ranges in objective, matrix and rhs
+        check_scaling(
+            self,
+            rtc_logger,
+            rtc_logs_list,
+        )
 
         pipes = solution.energy_system_components.get("heat_pipe")
         for pipe in pipes:
