+++ tests/modified_tests/test_multiple_in_and_out_port_components.py
@@ -7,6 +7,11 @@ from mesido.util import run_esdl_mesido_optimization
 
 import numpy as np
 
+from utils_test_scaling import (
+    check_scaling,
+    create_problem_with_debug_info,
+)
+
 from utils_tests import demand_matching_test, energy_conservation_test, heat_to_discharge_test
 
 
@@ -60,8 +65,14 @@ class TestHEX(TestCase):
         }
         # -----------------------------------------------------------------------------------------
 
+        (
+            heat_problem_post_scaling,
+            rtc_logger,
+            rtc_logs_list,
+        ) = create_problem_with_debug_info(HeatProblemPost)
+
         solution = run_esdl_mesido_optimization(
-            HeatProblemPost,
+            heat_problem_post_scaling,
             base_folder=base_folder,
             esdl_file_name="heat_exchanger.esdl",
             esdl_parser=ESDLFileParser,
@@ -83,6 +94,7 @@ class TestHEX(TestCase):
         demand_matching_test(solution, results)
         heat_to_discharge_test(solution, results)
         energy_conservation_test(solution, results)
+        check_scaling(self, rtc_logger, rtc_logs_list)
 
         np.testing.assert_allclose(prim_heat * eff, sec_heat)
 
@@ -154,8 +166,14 @@ class TestHP(TestCase):
         }
         # -----------------------------------------------------------------------------------------
 
+        (
+            heat_problem_post,
+            rtc_logger,
+            rtc_logs_list,
+        ) = create_problem_with_debug_info(HeatProblemPost)
+
         solution = run_esdl_mesido_optimization(
-            HeatProblemPost,
+            heat_problem_post,
             base_folder=base_folder,
             esdl_file_name="heat_pump.esdl",
             esdl_parser=ESDLFileParser,
@@ -182,6 +200,7 @@ class TestHP(TestCase):
         demand_matching_test(solution, results)
         heat_to_discharge_test(solution, results)
         energy_conservation_test(solution, results)
+        check_scaling(self, rtc_logger, rtc_logs_list)
 
         # We check the energy converted betweeen the commodities
         np.testing.assert_allclose(power_elec * parameters["GenericConversion_3d3f.COP"], sec_heat)
