+++ tests/modified_tests/test_varying_temperature.py
@@ -8,6 +8,11 @@ from mesido.util import run_esdl_mesido_optimization
 
 import numpy as np
 
+from utils_test_scaling import (
+    check_scaling,
+    create_problem_with_debug_info,
+)
+
 from utils_tests import demand_matching_test, energy_conservation_test, heat_to_discharge_test
 
 
@@ -32,8 +37,14 @@ class TestVaryingTemperature(TestCase):
 
         base_folder = Path(run_1a.__file__).resolve().parent.parent
 
+        (
+            heat_problem_t_var_scaling,
+            rtc_logger,
+            rtc_logs_list,
+        ) = create_problem_with_debug_info(HeatProblemTvar)
+
         heat_problem = run_esdl_mesido_optimization(
-            HeatProblemTvar,
+            heat_problem_t_var_scaling,
             base_folder=base_folder,
             esdl_file_name="1a.esdl",
             esdl_parser=ESDLFileParser,
@@ -59,6 +70,7 @@ class TestVaryingTemperature(TestCase):
         demand_matching_test(heat_problem, results)
         energy_conservation_test(heat_problem, results)
         heat_to_discharge_test(heat_problem, results)
+        check_scaling(self, rtc_logger, rtc_logs_list)
 
         parameters = heat_problem.parameters(0)
 
@@ -96,8 +108,14 @@ class TestVaryingTemperature(TestCase):
 
         base_folder = Path(run_3a.__file__).resolve().parent.parent
 
+        (
+            heat_problem_t_var_sup_scaling,
+            rtc_logger,
+            rtc_logs_list,
+        ) = create_problem_with_debug_info(HeatProblemTvarsup)
+
         heat_problem = run_esdl_mesido_optimization(
-            HeatProblemTvarsup,
+            heat_problem_t_var_sup_scaling,
             base_folder=base_folder,
             esdl_file_name="3a.esdl",
             esdl_parser=ESDLFileParser,
@@ -129,6 +147,7 @@ class TestVaryingTemperature(TestCase):
         demand_matching_test(heat_problem, results)
         energy_conservation_test(heat_problem, results)
         heat_to_discharge_test(heat_problem, results)
+        check_scaling(self, rtc_logger, rtc_logs_list)
 
         parameters = heat_problem.parameters(0)
 
@@ -168,8 +187,14 @@ class TestVaryingTemperature(TestCase):
 
         base_folder = Path(run_3a.__file__).resolve().parent.parent
 
+        (
+            heat_problem_t_var_ret_scaling,
+            rtc_logger,
+            rtc_logs_list,
+        ) = create_problem_with_debug_info(HeatProblemTvarret)
+
         heat_problem = run_esdl_mesido_optimization(
-            HeatProblemTvarret,
+            heat_problem_t_var_ret_scaling,
             base_folder=base_folder,
             esdl_file_name="3a.esdl",
             esdl_parser=ESDLFileParser,
@@ -194,6 +219,7 @@ class TestVaryingTemperature(TestCase):
         demand_matching_test(heat_problem, results)
         energy_conservation_test(heat_problem, results)
         heat_to_discharge_test(heat_problem, results)
+        check_scaling(self, rtc_logger, rtc_logs_list)
 
         parameters = heat_problem.parameters(0)
 
@@ -232,8 +258,14 @@ class TestVaryingTemperature(TestCase):
 
         base_folder = Path(run_heat_exchanger.__file__).resolve().parent.parent
 
+        (
+            heat_problem_t_var_scaling,
+            rtc_logger,
+            rtc_logs_list,
+        ) = create_problem_with_debug_info(HeatProblemTvar)
+
         heat_problem = run_esdl_mesido_optimization(
-            HeatProblemTvar,
+            heat_problem_t_var_scaling,
             base_folder=base_folder,
             esdl_file_name="heat_exchanger.esdl",
             esdl_parser=ESDLFileParser,
@@ -261,6 +293,7 @@ class TestVaryingTemperature(TestCase):
         demand_matching_test(heat_problem, results)
         energy_conservation_test(heat_problem, results)
         heat_to_discharge_test(heat_problem, results)
+        check_scaling(self, rtc_logger, rtc_logs_list)
 
         parameters = heat_problem.parameters(0)
 
@@ -298,8 +331,14 @@ class TestVaryingTemperature(TestCase):
 
         base_folder = Path(run_heat_exchanger.__file__).resolve().parent.parent
 
+        (
+            heat_problem_tvar_disable_hex_scaling,
+            rtc_logger,
+            rtc_logs_list,
+        ) = create_problem_with_debug_info(HeatProblemTvarDisableHEX)
+
         heat_problem = run_esdl_mesido_optimization(
-            HeatProblemTvarDisableHEX,
+            heat_problem_tvar_disable_hex_scaling,
             base_folder=base_folder,
             esdl_file_name="heat_exchanger.esdl",
             esdl_parser=ESDLFileParser,
@@ -323,6 +362,7 @@ class TestVaryingTemperature(TestCase):
         demand_matching_test(heat_problem, results)
         energy_conservation_test(heat_problem, results)
         heat_to_discharge_test(heat_problem, results)
+        check_scaling(self, rtc_logger, rtc_logs_list)
 
     def test_hex_temperature_variation_secondary(self):
         """
@@ -340,8 +380,14 @@ class TestVaryingTemperature(TestCase):
 
         base_folder = Path(run_heat_exchanger.__file__).resolve().parent.parent
 
+        (
+            heat_problem_tvar_secondary_scaling,
+            rtc_logger,
+            rtc_logs_list,
+        ) = create_problem_with_debug_info(HeatProblemTvarSecondary)
+
         heat_problem = run_esdl_mesido_optimization(
-            HeatProblemTvarSecondary,
+            heat_problem_tvar_secondary_scaling,
             base_folder=base_folder,
             esdl_file_name="heat_exchanger.esdl",
             esdl_parser=ESDLFileParser,
@@ -366,6 +412,7 @@ class TestVaryingTemperature(TestCase):
         demand_matching_test(heat_problem, results)
         energy_conservation_test(heat_problem, results)
         heat_to_discharge_test(heat_problem, results)
+        check_scaling(self, rtc_logger, rtc_logs_list)
 
     def test_heat_pump_varying_temperature(self):
         """
@@ -378,8 +425,14 @@ class TestVaryingTemperature(TestCase):
 
         base_folder = Path(run_heat_pump.__file__).resolve().parent.parent
 
+        (
+            heat_problem_tvar_scaling,
+            rtc_logger,
+            rtc_logs_list,
+        ) = create_problem_with_debug_info(HeatProblemTvar)
+
         heat_problem = run_esdl_mesido_optimization(
-            HeatProblemTvar,
+            heat_problem_tvar_scaling,
             base_folder=base_folder,
             esdl_file_name="heat_pump.esdl",
             esdl_parser=ESDLFileParser,
@@ -393,6 +446,7 @@ class TestVaryingTemperature(TestCase):
         demand_matching_test(heat_problem, results)
         energy_conservation_test(heat_problem, results)
         heat_to_discharge_test(heat_problem, results)
+        check_scaling(self, rtc_logger, rtc_logs_list)
 
         expected_cop = (
             parameters["GenericConversion_3d3f.efficiency"]
